新しいヘッダファイル`9cc.h`を作成し、トークンおよび抽象構文木の構造体を定義しました。また、コード生成を行う`codegen.c`とメイン関数を含む`main.c`を追加し、Makefileを修正してビルドプロセスを改善しました。`parse.c`を新規作成しましたが、内容は未実装です。
